+++ rmk-config/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Update toml config fields
+
 ## [0.1.6] - 2024-10-03
 
 - Fix no device on vial desktop